x/ml/types/message_ctrl_test.go
package types

import (
	"testing"

	"github.com/blockentropy/entropy/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCtrl_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCtrl
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCtrl{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCtrl{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
